requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Jinja2==2.7.2
MarkupSafe==0.23
PyJWT==0.2.3
PyYAML==3.11
Werkzeug==0.9.4
amqp==1.4.5
aniso8601==0.82
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.17
binaryornot==0.3.0
blessings==1.6
celery==3.1.11
click==3.3
cookiecutter==0.7.2
enum34==1.0
flake8==2.1.0
flask-restful-swagger==0.10
flex==3.3.0
gevent==1.0.1
gnureadline==6.3.3
greenlet==0.4.2
gunicorn==18.0
ijson==1.1
ipdb==0.8
ipython==2.1.0
iso8601==0.1.10
itsdangerous==0.24
kombu==3.0.16
mccabe==0.2.1
nose==1.3.3
numpy==1.8.1
pep8==1.5.6
pika==0.9.13
py==1.4.20
pyaml==14.05.7
pycrypto==2.6.1
pyflakes==0.8.1
pyparsing==2.0.2
pyswagger==0.8.4
pytest==2.5.2
python-dateutil==2.2
pytz==2014.3
pyzmq==14.3.0
redis==2.10.1
requests==2.5.3
rfc3987==1.3.4
six==1.9.0
Sphinx==1.2.3
sphinx-rtd-theme==0.1.6
tornado==3.2.1
tox==1.7.1
uWSGI==2.0.5.1
validate-email==1.2
virtualenv==1.11.6
wheel==0.24.0
wsgiref==0.1.2
